Adding  delete method to allow for deleting movies
I made changes to the main.py and movies.py files to add the DELETE method that follows similar formatting to the other methods. I ran it a few times to test and upon deletion, when entering the same unique id it will show an error instead of the previous result that would be shown.

This addresses issue #3