Adding a POST method to be able to post new movie entries
I made changes to main.py, in which I added the post method and changed the if statement in the get method. I tested out the new feature by POSTing a new movie and adding a cast. I then used the GET method and put id = 100 to go to the end of the list since it gets added on to the end and it was there.

This addresses issue #4.