Fetching a movie from fast API based on Movie ID
On the backend, I changed the main.py file to comment out the else print statement in my get method, and on the front end I added a bunch of things to fetch the results. I tested this by looking at the console while typing in different numbers to make sure the input was being registered.

This addresses issue #7